feat: update timestamp to milliseconds
BREAKING CHANGE: causes all timestamps to be invalid compared to
the previous format

+++ src/rpc/net.cpp
@@ -889,7 +889,7 @@ static RPCHelpMan getnodeaddresses()
 
     for (const CAddress& addr : vAddr) {
         UniValue obj(UniValue::VOBJ);
-        obj.pushKV("time", (int)addr.nTime);
+        obj.pushKV("time", (uint64_t)addr.nTime);
         obj.pushKV("services", (uint64_t)addr.nServices);
         obj.pushKV("address", addr.ToStringIP());
         obj.pushKV("port", addr.GetPort());
